+++ library/src/main/java/com/devbrackets/android/exomedia/core/api/VideoViewApi.java
@@ -62,6 +62,9 @@ public interface VideoViewApi {
 
     void pause();
 
+    /**
+     * Performs the functionality to stop the video in playback
+     */
     void stopPlayback();
 
     /**
